# openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI документация для фулфилмент сервиса
  version: v1
servers:
- url: http://localhost:8080
  description: Url сервера
...
paths:
...
  /api/v1/account/register:
    post:
      tags:
      - Account API
      summary: Регистрация нового пользователя
      description: |-
        Сохраняет нового пользователя и информацию о товарах, полученную от Wildberries.
        Если api ключ недействителен, пользователь не будет зарегистрирован
      operationId: register
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      requestBody:
        description: Данные для регистрации
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationRequest"
        required: true
      responses:
        "201":
          description: Успешное сохранение
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "400":
          description: "Некорректный запрос, в том числе недействительный api ключ\
            \ Wildberries"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:04:22.00Z
                status: 400
                error: "Wildberries API Error: unauthorized"
                message: "token problem; token is malformed: token contains an invalid\
                  \ number of segments"
                path: /api/v1/account/register
        "401":
          description: Пользователь не аутентифицирован через telegram
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:28:24.68Z
                status: 401
                error: Telegram authentication failed
                message: Init data expired. Log in again via Telegram
                path: /api/v1/account/register
        "409":
          description: "Пользователь с таким именем, email, api ключом или телегра\
            ммом уже существует"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:16:24.64Z
                status: 409
                error: Resource conflict
                message: Account with this wb api key already exists
                path: /api/v1/account/register
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:22:31.00Z
                status: 500
                error: Database error
                message: Couldn't save account
                path: /api/v1/account/register
        "502":
          description: Ошибка при работе с api Wildberries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:22:31.00Z
                status: 502
                error: "Wildberries API Error: method not allowed"
                message: allowed methods are listed in the Allow header (POST)
                path: /api/v1/account/register
      security:
      - TelegramAuth: []
...
  /api/v1/account/wb-api-key:
    get:
      tags:
      - Account API
      summary: Получение api ключа пользователя
      description: Возвращает api ключ пользователя по telegram id из initData
      operationId: getWbApiKey
      responses:
        "200":
          description: Успешно
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyResponse"
        "401":
          description: Пользователь не аутентифицирован через telegram
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:28:24.68Z
                status: 401
                error: Telegram authentication failed
                message: Init data expired. Log in again via Telegram
                path: /api/v1/account/wb-api-key
        "404":
          description: Api ключ не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:04:22.00Z
                status: 404
                error: No data found
                message: Wb api key was not found for this account
                path: /api/v1/account/wb-api-key
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:22:31.00Z
                status: 500
                error: Database error
                message: Couldn't get wb api key
                path: /api/v1/account/wb-api-key
      security:
      - TelegramAuth: []
    patch:
      tags:
      - Account API
      summary: Обновление api ключа пользователя
      description: Обновляет api ключ пользователя. Корректность нового ключа не проверяется
      operationId: changeWbApiKey
      requestBody:
        description: Новое значение ключа
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeKeyRequest"
        required: true
      responses:
        "200":
          description: Успешное обновление
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangeKeyResponse"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T14:33:51.19Z
                status: 400
                error: Validation error
                message: Incorrect data was entered
                path: /api/v1/account/wb-api-key
                details:
                  wbApiKey: The wb api key cannot be empty
        "401":
          description: Пользователь не аутентифицирован через telegram
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:28:24.68Z
                status: 401
                error: Telegram authentication failed
                message: Init data expired. Log in again via Telegram
                path: /api/v1/account/wb-api-key
        "404":
          description: Аккаунт для обновления api ключа не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:04:22.00Z
                status: 404
                error: No data found
                message: Account not found
                path: /api/v1/account/wb-api-key
        "409":
          description: Пользователь с таким api ключом уже существует
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T14:30:23.97Z
                status: 409
                error: Resource conflict
                message: "It is not possible to save this wb api key, it already exists\
                  \ in the database"
                path: /api/v1/account/wb-api-key
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:22:31.00Z
                status: 500
                error: Database error
                message: Couldn't change the wb api key
                path: /api/v1/account/wb-api-key
      security:
      - TelegramAuth: []
...
components:
  schemas:
...
    AppErrorResponse:
      type: object
      description: Стандартный ответ ошибки
      properties:
        timestamp:
          type: string
          description: "Дата и время, когда произошла ошибка"
          example: 2025-07-28T11:51:39.39Z
        status:
          type: integer
          format: int32
          description: Статусный код HTTP
          example: 400
        error:
          type: string
          description: Заголовок ошибки
          example: Validation error
        message:
          type: string
          description: Описание ошибки
          example: Incorrect data was entered
        path:
          type: string
          description: "Путь запроса, где произошла ошибка"
          example: /api/v1/shipment
        details:
          type: object
          additionalProperties:
            type: string
          description: Дополнительные сведения при ошибках валидации
          example:
            quantity: The number must be greater than 0
            warehouseId: The warehouse id cannot be empty
            warehouseName: The warehouse name cannot be empty
...
    AccountResponse:
      type: object
      description: Ответ с информацией о пользователе
      properties:
        id:
          type: integer
          format: int64
          description: id пользователя
          example: 100050
        name:
          type: string
          description: Имя пользователя
          example: Anna
        phone:
          type: string
          description: Номер телефона пользователя
          example: "+79123456789"
        email:
          type: string
          description: Email пользователя
          example: anna@example.com
    RegistrationRequest:
      type: object
      description: Запрос на регистрацию нового пользователя
      properties:
        name:
          type: string
          description: Уникальное имя пользователя
          example: Anna
          maxLength: 30
          minLength: 2
        phone:
          type: string
          description: Номер телефона пользователя
          example: "+79123456789"
          minLength: 1
          pattern: "^\\+?[1-9][0-9]{7,14}$"
        email:
          type: string
          description: Уникальный email пользователя
          example: anna@example.com
          maxLength: 255
          minLength: 0
        wbApiKey:
          type: string
          description: Уникальный api ключ Wildberries
          example: xYjJ9bGci0iJFUzI2NyIs1mtpZCI6I3jIwMjYwNTI2d...
          maxLength: 500
          minLength: 0
      required:
      - email
      - name
      - phone
      - wbApiKey
...
    KeyResponse:
      type: object
      description: Ответ с значением api ключа
      properties:
        wbApiKey:
          type: string
          description: Значение api ключа
          example: NC1hZTVjLTcwZDYtODBjMy1kMTA1...
  securitySchemes:
    TelegramAuth:
      type: apiKey
      description: Telegram WebApp initData
      name: Authorization
      in: header
    ApiSecretAuth:
      type: apiKey
      description: Секретный ключ API
      name: X-API-SECRET
      in: header